requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
